despachos

+++ app/Http/Controllers/Bodega/BodEgresosController.php
@@ -7,6 +7,7 @@ use App\Http\Controllers\Hacienda\InventarioEmpleadoController;
 use App\Models\Bodega\Bodega;
 use App\Models\Bodega\EgresoBodega;
 use App\Models\Bodega\EgresoBodegaDetalle;
+use App\Models\Bodega\EgresoBodegaDetalleTransfer;
 use App\Models\Hacienda\InventarioEmpleado;
 use Carbon\Carbon;
 use Illuminate\Http\Request;
@@ -35,54 +36,72 @@ class BodEgresosController extends Controller
             $params_array = json_decode($json, true);
             if (!empty($params_array) && count($params_array) > 0) {
                 DB::beginTransaction();
-                $validacion = $this->validationModel($params_array);
+                $validacion = Validator::make($params_array, [
+                    'cabecera.idempleado' => 'required',
+                    'cabecera.hacienda' => 'required',
+                ], [
+                    'cabecera.empleado.id.required' => "El empleado es necesario",
+                    'cabecera.hacienda.required' => "La hacienda es necesaria",
+                ]);;
 
                 if (!$validacion->fails()) {
                     //Procesar la cabecera
                     $cabecera = $params_array['cabecera'];
-                    //Convertir fecha
-                    $timestamp = strtotime(str_replace('/', '-', $cabecera['fecha']));
-                    $cabecera['fecha'] = date(config('constants.date'), $timestamp);
+                    $despachos = $params_array['detalle'];
+                    $transferencias = $params_array['transferencias'];
 
-                    $egreso = new EgresoBodega();
-                    $egreso->fecha_apertura = $cabecera['fecha'];
-                    $egreso->idempleado = $cabecera['idempleado'];
-                    $egreso->parcial = false;
-                    $egreso->final = false;
-                    $egreso->created_at = Carbon::now()->format(config('constants.format_date'));
-                    $egreso->updated_at = Carbon::now()->format(config('constants.format_date'));
-                    $egreso->estado = true;
+                    if (count($despachos) > 0 || count($transferencias) > 0) {
+                        //Convertir fecha
+                        $timestamp = strtotime(str_replace('/', '-', $cabecera['fecha']));
+                        $cabecera['fecha'] = date(config('constants.date'), $timestamp);
 
-                    if (!EgresoBodega::existe($egreso)) {
-                        $egreso->save();
-                        //Procesar los materiales despachados
-                        $despachos = $params_array['detalle'];
+                        $egreso = new EgresoBodega();
+                        $egreso->fecha_apertura = $cabecera['fecha'];
+                        $egreso->idempleado = $cabecera['idempleado'];
+                        $egreso->parcial = false;
+                        $egreso->final = false;
+                        $egreso->created_at = Carbon::now()->format(config('constants.format_date'));
+                        $egreso->updated_at = Carbon::now()->format(config('constants.format_date'));
+                        $egreso->estado = true;
 
-                        foreach ($despachos as $despacho) {
-                            $timestamp = strtotime(str_replace('/', '-', $despacho['time']));
-                            $despacho['time'] = date(config('constants.date'), $timestamp);
+                        if (!EgresoBodega::existe($egreso)) {
+                            $egreso->save();
+                            //Procesar los materiales despachados
+                            foreach ($despachos as $despacho) {
+                                $timestamp = strtotime(str_replace('/', '-', $despacho['time']));
+                                $despacho['time'] = date(config('constants.date'), $timestamp);
 
-                            $egreso_detalle = new EgresoBodegaDetalle();
-                            $egreso_detalle->idegreso = $egreso->id;
-                            $egreso_detalle->idmaterial = $despacho['idmaterial'];
-                            $egreso_detalle->fecha_salida = $despacho['time'];
-                            $egreso_detalle->cantidad = $despacho['cantidad'];
-                            $egreso_detalle->created_at = Carbon::now()->format(config('constants.format_date'));
-                            $egreso_detalle->updated_at = Carbon::now()->format(config('constants.format_date'));
-                            $egreso_detalle->estado = true;
-                            $egreso_detalle->save();
+                                $egreso_detalle = new EgresoBodegaDetalle();
+                                $egreso_detalle->idegreso = $egreso->id;
+                                $egreso_detalle->idmaterial = $despacho['idmaterial'];
+                                $egreso_detalle->fecha_salida = $despacho['time'];
+                                $egreso_detalle->cantidad = $despacho['cantidad'];
+                                $egreso_detalle->created_at = Carbon::now()->format(config('constants.format_date'));
+                                $egreso_detalle->updated_at = Carbon::now()->format(config('constants.format_date'));
+                                $egreso_detalle->estado = true;
+                                $egreso_detalle->save();
 
-                            //Procesar Inventario de empleado
-                            InventarioEmpleadoController::storeInventario($egreso->idempleado, $egreso_detalle);
-                            //return response()->json($inventario);
+                                //Procesar Inventario de empleado
+                                InventarioEmpleadoController::storeInventario($egreso->idempleado, $egreso_detalle);
+                                //return response()->json($inventario);
+                            }
+
+                            $this->out = $this->respuesta_json('success', 200, 'Despacho registrado con éxito!!!');
+
+                            //Procesar transferencias de saldo en caso de existir
+                            if (!empty($transferencias) && count($transferencias) > 0) {
+                                //Proceso de transferencias
+                                $resultado_transferencias = $this->SaldoTransfer($transferencias, $egreso);
+                                $this->out['transferencias'] = $resultado_transferencias;
+                            }
+
+                            DB::commit();
+                        } else {
+                            throw new \Exception('Ya se ha registrado un egreso para este empleado.');
                         }
-
-                        DB::commit();
-                        $this->out = $this->respuesta_json('success', 200, 'Despacho registrado con éxito!!!');
                     } else {
-                        throw new \Exception('Ya se ha registrado un egreso para este empleado.');
+                        throw new \Exception('No se puede procesar la transacción.');
                     }
-
                 } else {
                     $this->out['validacion'] = $validacion->errors()->all();
                     throw new \Exception('Error en la validacion de datos.');
@@ -108,7 +127,7 @@ class BodEgresosController extends Controller
     {
         try {
             $fecha = $request->get('fecha', null);
-            if (!empty($fecha) && !is_null($fecha)) {
+            if (!empty($fecha)) {
                 $existe = EgresoBodega::existeByEmpleado($idempleado, $fecha);
                 if (is_object($existe)) {
                     $egreso = EgresoBodega::from("BOD_EGRESOS as egreso")
@@ -121,7 +140,7 @@ class BodEgresosController extends Controller
                             }]);
                         }])
                         ->with(['egresoDetalle' => function ($query) {
-                            $query->select('id', 'idegreso', 'fecha_salida as fecha', 'idmaterial', 'cantidad');
+                            $query->select('id', 'idegreso', 'fecha_salida as fecha', 'idmaterial', 'cantidad', 'movimiento');
                             $query->with(['materialdetalle' => function ($query) {
                                 $query->select('id', 'codigo', 'descripcion', 'stock');
                             }]);
@@ -135,9 +154,10 @@ class BodEgresosController extends Controller
                 } else {
                     throw new \Exception('No se encontraron despachos para esta semana.');
                 }
-
-
+            } else {
+                throw new \Exception('No se ha recibido la fecha.');
             }
+
         } catch (\Exception $ex) {
             $this->out['error'] = $ex->getMessage();
         }
@@ -152,7 +172,7 @@ class BodEgresosController extends Controller
             $params_array = json_decode($json, true);
             $egreso = EgresoBodega::existeById($id);
 
-            if (!empty($id) && !is_null($id)) {
+            if (!empty($id)) {
                 if (!empty($params_array) && count($params_array) > 0) {
                     if (is_object($egreso)) {
                         //Solo atualizamos el detalle
@@ -197,40 +217,43 @@ class BodEgresosController extends Controller
 
                                 $existe = EgresoBodegaDetalle::existe($egreso_detalle);
                                 if (is_object($existe)) {
-                                    $egreso_detalle = $existe;
+                                    //Verificar si no es un detalle por transferencia
+                                    $is_Transferencia = EgresoBodegaDetalleTransfer::existTransferbyDetalleEgreso($existe->id);
+                                    if (!is_object($is_Transferencia)) {
+                                        $egreso_detalle = $existe;
 
-                                    //Consultar si se ha hecho una transferencia de este egreso.
+                                        //Consultar si se ha hecho una transferencia de este egreso.
+                                        if (isset($despacho['delete']) && filter_var($despacho['delete'], FILTER_VALIDATE_BOOL)) {
+                                            //Eliminar item.
 
-                                    if (isset($despacho['delete']) && filter_var($despacho['delete'], FILTER_VALIDATE_BOOL)) {
-                                        //Eliminar item.
+                                            $respuesta = InventarioEmpleadoController::reduceInventario($egreso->idempleado, $egreso_detalle);
 
-                                        $respuesta = InventarioEmpleadoController::reduceInventario($egreso->idempleado, $egreso_detalle);
+                                            if ($respuesta['negativo']) {
+                                                InventarioEmpleadoController::storeInventario($egreso->idempleado, $egreso_detalle, true);
+                                                $romper_ciclo_detalle = "Un registro no se pudo eliminar, el consumo no puede ser mayor a lo que se despacho.";
+                                                break;
+                                            }
 
-                                        if ($respuesta['negativo']) {
-                                            InventarioEmpleadoController::storeInventario($egreso->idempleado, $egreso_detalle, true);
-                                            $romper_ciclo_detalle = "Un registro no se pudo eliminar, el consumo no puede ser mayor a lo que se despacho.";
-                                            break;
+                                            $egreso_detalle->delete();
+
+                                        } else {
+                                            //Editar item
+                                            $cantidad_a_saldar = $egreso_detalle->cantidad;
+                                            $egreso_detalle->cantidad = $despacho['cantidad'];
+
+                                            //Actualizacion de los items de despacho
+                                            //Procesar Inventario de empleado
+                                            $respuesta = InventarioEmpleadoController::storeInventario($egreso->idempleado,
+                                                $egreso_detalle, false, $cantidad_a_saldar);
+
+                                            if ($respuesta['negativo']) {
+                                                $romper_ciclo_detalle = "Un registro no se pudo editar, el consumo no puede ser mayor a lo que se despacho.";
+                                                break;
+                                            }
+
+                                            $egreso_detalle->updated_at = Carbon::now()->format(config('constants.format_date'));
+                                            $egreso_detalle->save();
                                         }
-
-                                        $egreso_detalle->delete();
-
-                                    } else {
-                                        //Editar item
-                                        $cantidad_a_saldar = $egreso_detalle->cantidad;
-                                        $egreso_detalle->cantidad = $despacho['cantidad'];
-
-                                        //Actualizacion de los items de despacho
-                                        //Procesar Inventario de empleado
-                                        $respuesta = InventarioEmpleadoController::storeInventario($egreso->idempleado,
-                                            $egreso_detalle, false, $cantidad_a_saldar);
-
-                                        if ($respuesta['negativo']) {
-                                            $romper_ciclo_detalle = "Un registro no se pudo editar, el consumo no puede ser mayor a lo que se despacho.";
-                                            break;
-                                        }
-
-                                        $egreso_detalle->updated_at = Carbon::now()->format(config('constants.format_date'));
-                                        $egreso_detalle->save();
                                     }
                                 } else {
                                     //Edicion de transaccion, pero se añadio un nuevo item de despacho
@@ -247,6 +270,14 @@ class BodEgresosController extends Controller
 
                             $this->out = $this->respuesta_json('success', 200, 'Despacho actualizado con éxito!!!');
                             $this->out['error_ciclo'] = $romper_ciclo_detalle;
+
+                            //Procesar transferencias de saldo en caso de existir
+                            $transferencias = $params_array['transferencias'];
+                            if (!empty($transferencias) && count($transferencias) > 0) {
+                                //Proceso de transferencias
+                                $resultado_transferencias = $this->SaldoTransfer($transferencias, $egreso);
+                                $this->out['transferencias'] = $resultado_transferencias;
+                            }
 
                             DB::commit();
                         } else {
@@ -347,6 +378,88 @@ class BodEgresosController extends Controller
         }
 
         return response()->json($this->out, $this->out['code']);
+    }
+
+    public function modelTransferenciatoEgreso($data, $id_cabecera_egreso, $movimiento = 'CREDIT-SLD') //ACREDITACION DE SALDO
+    {
+        $timestamp = strtotime(str_replace('/', '-', $data['time']));
+        $data['time'] = date(config('constants.date'), $timestamp);
+
+        $egreso_detalle = new EgresoBodegaDetalle();
+        $egreso_detalle->idegreso = $id_cabecera_egreso;
+        $egreso_detalle->idmaterial = $data['idmaterial'];
+        $egreso_detalle->fecha_salida = $data['time'];
+        $egreso_detalle->cantidad = $data['cantidad'];
+        $egreso_detalle->movimiento = $movimiento;
+        $egreso_detalle->created_at = Carbon::now()->format(config('constants.format_date'));
+        $egreso_detalle->updated_at = Carbon::now()->format(config('constants.format_date'));
+        $egreso_detalle->estado = true;
+
+        return $egreso_detalle;
+    }
+
+    public function SaldoTransfer($transferencias, $egreso)
+    {
+        $save = false;
+        $respuestas = [];
+
+        foreach ($transferencias as $transferencia) {
+            //Guardamos en la tabla de transferencias
+            //Registramos el debito al empleado
+            $transferencia_credito = $this->modelTransferenciatoEgreso($transferencia, $egreso->id); //CREDITO
+            $existe = EgresoBodegaDetalle::existe($transferencia_credito);
+            if (is_object($existe)) {
+                $transferencia_credito = $existe;
+                $cantidad_saldar = $transferencia_credito->cantidad;
+                $transferencia_credito->cantidad += $transferencia['cantidad'];
+                $save = InventarioEmpleadoController::updateInventarioByTransferSaldo($egreso->idempleado, $transferencia_credito, $cantidad_saldar)['status'];
+            } else {
+                $save = InventarioEmpleadoController::updateInventarioByTransferSaldo($egreso->idempleado, $transferencia_credito)['status'];
+            }
+
+
+            if ($save) {
+                $save = $transferencia_credito->save();
+                if ($save) {
+                    //Guardamos la transferrencia
+                    $saldo_inv_transfer = new EgresoBodegaDetalleTransfer();
+                    $saldo_inv_transfer->idEgreso = $transferencia_credito->id;
+                    $saldo_inv_transfer->idInvEmp = $transferencia['idInv'];
+                    $saldo_inv_transfer->cantidad = $transferencia['cantidad'];
+                    $saldo_inv_transfer->debito = true;
+                    $saldo_inv_transfer->created_at = Carbon::now()->format(config('constants.format_date'));
+
+                    //Procedemos al debito de los inventarios de donde se ha tomado el saldo
+                    $inventario = InventarioEmpleado::where(['id' => $transferencia['idInv']])->first();
+                    $existe_transfer = EgresoBodegaDetalleTransfer::existTransfer($saldo_inv_transfer);
+
+                    if (is_object($existe_transfer)) {
+                        $saldo_inv_transfer = $existe_transfer;
+                        //Reduce el inventario
+                        $inventario->tot_devolucion += $transferencia['cantidad'];
+                        //Edita el registro
+                        $saldo_inv_transfer->cantidad += $transferencia['cantidad'];
+                    } else {
+                        $inventario->tot_devolucion = $transferencia['cantidad'];
+                    }
+
+                    $inventario->sld_final = ($inventario->sld_inicial + $inventario->tot_egreso) - ($inventario->tot_consumo + $inventario->tot_devolucion);
+                    $save = $inventario->save();
+
+                    if ($save) {
+                        $saldo_inv_transfer->updated_at = Carbon::now()->format(config('constants.format_date'));
+                        $save = $saldo_inv_transfer->save();
+                    }
+                }
+
+                array_push($respuestas, [
+                    'save' => $save,
+                    'transferencia' => $transferencia
+                ]);
+            }
+        }
+
+        return $respuestas;
     }
 
     public function validationModel($params_array)
